feat: 랭킹 서비스 추상화 적용
